+++ learning_assistant/constants.py
@@ -1,0 +1,10 @@
+"""
+Constants for the learning_assistant app.
+"""
+# Pulled from edx-platform. Will correctly capture both old- and new-style
+# course ID strings.
+INTERNAL_COURSE_KEY_PATTERN = r'([^/+]+(/|\+)[^/+]+(/|\+)[^/?]+)'
+
+EXTERNAL_COURSE_KEY_PATTERN = r'([A-Za-z0-9-_:]+)'
+
+COURSE_ID_PATTERN = rf'(?P<course_id>({INTERNAL_COURSE_KEY_PATTERN}|{EXTERNAL_COURSE_KEY_PATTERN}))'